fix for reading a zero byte file.  Also adding fixes from http://www.beowulf-underground.org/pipermail/pvfs2-developers/2006-January001741.html and http://www.beowulf-underground.org/pipermail/pvfs2-developers/2005-December/001737.html
SVN revision 4826

+++ src/common/misc/pint-perf-counter.c
@@ -171,7 +171,7 @@ void PINT_perf_reset(
 
     /* set initial timestamp */
     gettimeofday(&tv, NULL);
-    pc->start_time_array_ms[0] = tv.tv_sec*1000 +
+    pc->start_time_array_ms[0] = ((uint64_t)tv.tv_sec)*1000 +
 	tv.tv_usec/1000;
 
     gen_mutex_unlock(pc->mutex);
